+++ src/lightning_app/cli/lightning_cli.py
@@ -304,6 +304,7 @@ def _run_app(
     blocking: bool,
     open_ui: bool,
     env: tuple,
+    secret: tuple,
 ) -> None:
     file = _prepare_file(file)
 
@@ -319,11 +320,18 @@ def _run_app(
                 "Caching is a property of apps running in cloud. "
                 "Using the flag --no-cache in local execution is not supported."
             )
+        if secret:
+            raise click.ClickException(
+                "Secrets can only be used for apps running in cloud. "
+                "Using the option --secret in local execution is not supported."
+            )
 
     env_vars = _format_input_env_variables(env)
     os.environ.update(env_vars)
 
-    def on_before_run(*args: Any) -> None:
+    secrets = _format_input_env_variables(secret)
+
+    def on_before_run(*args, **kwargs) -> None:
         if open_ui and not without_server:
             click.launch(get_app_url(runtime_type, *args))
 
@@ -341,6 +349,7 @@ def _run_app(
         on_before_run=on_before_run,
         name=name,
         env_vars=env_vars,
+        secrets=secrets,
         cluster_id=cluster_id,
     )
     if runtime_type == RuntimeType.CLOUD:
@@ -364,8 +373,14 @@ def run() -> None:
     "--no-cache", is_flag=True, default=False, help="Disable caching of packages " "installed from requirements.txt"
 )
 @click.option("--blocking", "blocking", type=bool, default=False)
-@click.option("--open-ui", type=bool, default=True, help="Decide whether to launch the app UI in a web browser")
-@click.option("--env", type=str, default=[], multiple=True, help="Env variables to be set for the app.")
+@click.option(
+    "--open-ui",
+    type=bool,
+    default=True,
+    help="Decide whether to launch the app UI in a web browser"
+)
+@click.option("--env", type=str, default=[], multiple=True, help="Environment variables to be set for the app.")
+@click.option("--secret", type=str, default=[], multiple=True, help="Secret variables to be set for the app.")
 @click.option("--app_args", type=str, default=[], multiple=True, help="Collection of arguments for the app.")
 def run_app(
     file: str,
@@ -377,10 +392,11 @@ def run_app(
     blocking: bool,
     open_ui: bool,
     env: tuple,
+    secret: tuple,
     app_args: tuple,
 ) -> None:
     """Run an app from a file."""
-    _run_app(file, cloud, cluster_id, without_server, no_cache, name, blocking, open_ui, env)
+    _run_app(file, cloud, cluster_id, without_server, no_cache, name, blocking, open_ui, env, secret)
 
 
 @_main.group(hidden=True)